@@ -1,7 +1,8 @@
 package gt.com.diego.wallapp;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class CreatePostActivity extends AppCompatActivity {
 
